# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Calendar API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/hosts/{handle}": {
      "get": {
        "tags": [
          "hosts"
        ],
        "summary": "Get Public Host",
        "description": "A Host's Public name and the Event Types Guests can book.",
        "operationId": "get_public_host",
        "parameters": [
          {
            "name": "handle",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Handle"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicHost"
                }
              }
            }
          },
          "404": {
            "description": "No Host with this Handle"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/hosts/{handle}/event-types/{event_type_id}": {
      "get": {
        "tags": [
          "hosts"
        ],
        "summary": "Get Public Event Type",
        "description": "An Event Type a Guest can book, with its Host's Public name.",
        "operationId": "get_public_event_type",
        "parameters": [
          {
            "name": "handle",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Handle"
            }
          },
          {
            "name": "event_type_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Event Type Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicEventTypeDetail"
                }
              }
            }
          },
          "404": {
            "description": "No such Host or bookable Event Type"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/hosts/{handle}/event-types/{event_type_id}/slots": {
      "get": {
        "tags": [
          "hosts"
        ],
        "summary": "List Slots",
        "description": "Free Slot start times (UTC) in [from, to), clipped to the booking window.",
        "operationId": "list_slots",
        "parameters": [
          {
            "name": "handle",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Handle"
            }
          },
          {
            "name": "event_type_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Event Type Id"
            }
          },
          {
            "name": "from",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time",
              "title": "From"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time",
              "title": "To"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Slots"
                }
              }
            }
          },
          "404": {
            "description": "No such Host or bookable Event Type"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/hosts/{handle}/event-types/{event_type_id}/bookings": {
      "post": {
        "tags": [
          "bookings"
        ],
        "summary": "Create Booking",
        "description": "Book a Slot as a Guest. No account needed (ADR 0002).",
        "operationId": "create_booking",
        "parameters": [
          {
            "name": "handle",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Handle"
            }
          },
          {
            "name": "event_type_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Event Type Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookingRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingConfirmation"
                }
              }
            }
          },
          "404": {
            "description": "No such Host or bookable Event Type"
          },
          "409": {
            "description": "The time was just taken by another Booking"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "BookingConfirmation": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "start": {
            "type": "string",
            "format": "date-time",
            "title": "Start"
          },
          "end": {
            "type": "string",
            "format": "date-time",
            "title": "End"
          },
          "event_type_title": {
            "type": "string",
            "title": "Event Type Title"
          },
          "host_public_name": {
            "type": "string",
            "title": "Host Public Name"
          },
          "guest_name": {
            "type": "string",
            "title": "Guest Name"
          },
          "guest_email": {
            "type": "string",
            "title": "Guest Email"
          },
          "guest_note": {
            "type": "string",
            "title": "Guest Note"
          }
        },
        "type": "object",
        "required": [
          "id",
          "start",
          "end",
          "event_type_title",
          "host_public_name",
          "guest_name",
          "guest_email",
          "guest_note"
        ],
        "title": "BookingConfirmation"
      },
      "BookingRequest": {
        "properties": {
          "start": {
            "type": "string",
            "format": "date-time",
            "title": "Start"
          },
          "guest_name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Guest Name"
          },
          "guest_email": {
            "type": "string",
            "format": "email",
            "title": "Guest Email"
          },
          "guest_note": {
            "type": "string",
            "maxLength": 2000,
            "title": "Guest Note",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "start",
          "guest_name",
          "guest_email"
        ],
        "title": "BookingRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PublicEventType": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "duration_minutes": {
            "type": "integer",
            "title": "Duration Minutes"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "description",
          "duration_minutes"
        ],
        "title": "PublicEventType"
      },
      "PublicEventTypeDetail": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "duration_minutes": {
            "type": "integer",
            "title": "Duration Minutes"
          },
          "host_public_name": {
            "type": "string",
            "title": "Host Public Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "description",
          "duration_minutes",
          "host_public_name"
        ],
        "title": "PublicEventTypeDetail"
      },
      "PublicHost": {
        "properties": {
          "public_name": {
            "type": "string",
            "title": "Public Name"
          },
          "handle": {
            "type": "string",
            "title": "Handle"
          },
          "event_types": {
            "items": {
              "$ref": "#/components/schemas/PublicEventType"
            },
            "type": "array",
            "title": "Event Types"
          }
        },
        "type": "object",
        "required": [
          "public_name",
          "handle",
          "event_types"
        ],
        "title": "PublicHost"
      },
...
      "Slots": {
        "properties": {
          "slots": {
            "items": {
              "type": "string",
              "format": "date-time"
            },
            "type": "array",
            "title": "Slots"
          }
        },
        "type": "object",
        "required": [
          "slots"
        ],
        "title": "Slots"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}